// src/visual-dashboard-frontend/src/interfaces/schemasReducer.ts
export interface AnchorPoint {
    id: number,
    x_coord: number,
    y_coord: number
}

interface MachineImgThumb {
    large: string
    large_height: number
    large_width: number
    small: string
    small_height: number
    small_width: number
}

export interface MachineSensorPoint {
    id: number
    x_coord: number
    y_coord: number
    side: string
    type_of_machine: number
}

export interface MachineType {
    anchor_points: AnchorPoint[]
    id: number,
    left_side_img: string,
    left_side_img_thumbs: MachineImgThumb,
    right_side_img: string,
    right_side_img_thumbs: MachineImgThumb,
    sensor_points_left: MachineSensorPoint[]
    sensor_points_right: MachineSensorPoint[]
    name: string
}

export interface Machine {
    x_coord: number,
    y_coord: number,
    scale: 1.0,
    type_of_machine: number,
    sensors?: Sensor
    flip: true,
    frontend_id?: number,
    id?: number
}

export interface Sensor {
    id: string,
    name: string,
    device: any,
    liveData?: any
    group_name: string
}

export interface SensorObj {
    sensor: Sensor,
    sensor_point: number,
}

export interface SchemaInMachineDetails {
    id: number,
    name: string,
    location: Location
}

export interface MachineExtended {
    x_coord: number,
    y_coord: number,
    scale: 1.0,
    type_of_machine: MachineType,
    sensors?: SensorObj[]
    flip: boolean,
    frontend_id: number,
    id?: number,
    name?: string,
    schema?: SchemaInMachineDetails
    schema_tab?: TabInList,
    selectedSensorId?: string,
    device_ids?: number[],
    system_state_sensor?: Sensor
}
// ...
export interface TabInList {
    id: number
    name: string,
    sensors_ids: any[]
}
// ...
export interface Location {
    id: number,
    title: string,
    published_schema?: PublishedSchema
    devices: any
}